+++ src/components/modal/AddRestaurantFormModal/index.js
@@ -1,3 +1,4 @@
+import "./index.css";
 class AddRestaurantFormModal {
   $target;
