+++ src/components/restaurants/RestaurantList/index.css
@@ -1,0 +1,7 @@
+.restaurant-list-container {
+  display: flex;
+  flex-direction: column;
+
+  padding: 0 16px;
+  margin: 16px 0;
+}
